+++ app/src/main/java/com/raudonikiss/weatherforecast/network/response/Sys.kt
@@ -4,7 +4,7 @@ data class Sys(
     val country: String,
     val id: Int,
     val message: Double,
-    val sunrise: Int,
-    val sunset: Int,
-    val type: Int
+    val sunrise: Long,
+    val sunset: Long,
+    val type: Double
 )
